Change from an after-change-function to an idle-timer
This is an attempt to stop larger buffers from lagging due to the
rather expensive full revaluation happening on every keystroke, which
becomes very noticeable very quickly.

This change introduces a new customisation variable,
literate-calc-mode-idle-time, which controls how long to wait after
editing to actually recalculate. It defaults to 1 second, which feels
like a reasonable value. A value of 0 should work just like the
previous implementation.

The implementation is a bit weird because run-with-idle-timer is
inherently global. We avoid running many more recalculations than
required by scheduling the recalculation on an edit, explicitly for
the current buffer. If there is already a timer set because we're in
the middle of typing, we cancel the existing timer so that there is at
most one timer per buffer. This means that buffers also recalculate
correctly if we switch away from them before the timer actually fires.

Because several edits in quick succession could cancel each others
timers, we recalculate the whole buffer once editing ceases, as we
cannot rely on the change information provided in the hook (unless we
were to aggregate them).