Only cancel the idle timer on exit if it exists
If activating the major or minor mode for the first time in the
buffer, the buffer-local idle timer variable has not been set. It will
be set by the after-change hook on the first edit. This means entering
and exiting the mode in a fresh buffer without any edits broke,
because the timer was nil, so it could not be cancelled.

The workaround is simple and self-explanatory.

Fixes #20