Enable digit grouping
This fixes #24.

Turns out calc has a whole set of really interesting options that
control formatting, precision, and so on.

A slight complication was that we store named results in the
buffer-local namespace as strings, and grouped results are interpreted
as lists, so this adds some slight overhead because we have to do the
calculation again with default options to get a raw result for future
use, but that should be fine.