fix #18: add an option to limit the buffer size for the minor mode
Sorry this took me so long @x61, I've been busy/procrastinating a lot.

If the option is set to a non-zero value, and the buffer size exceeds
that, minor mode startup is just skipped. This should be useful for
folks running the minor mode on a major-mode hook, such as org-mode.

The idea here being as well that for the major mode there's nothing
reasonable to fall back on, and if you're starting the major mode you
probably want the major mode. In the case of the minor mode you just
probably just want to avoid everything grinding to a halt on those
larger files where you actually don't care about LCM.