Avoid clobbering calc functions during replacement
This taught me a lot about how regexp matching works internally. The
process here is relatively simple: instead of blindly replacing all
variables with their respective values, we walk over the string to the
first match, check if it happens inside a reserved function name, if
not replace it, and keep walking until we hit the end of the string.
Repeat this for all variables in scope for each line.

Fixes #15, #31