Only block variable expansion for longer overlapping function names
The fix for #15 blocked variable expansion for cases where a variable
name would overlap with a builtin function. This works, but introduced
a new problem when a variable name is overlapped by a shorter builtin
function, such as `year`.

With this change the longer name will take precedence, with builtin
functions taking precedence for equal length.

Fixes #36.